+++ build_runner_core/test/generate/build_test.dart
@@ -243,22 +243,19 @@ void main() {
       });
 
       test('with placeholder as input', () async {
-        await testPhases(
-          [
-            applyToRoot(
-              PlaceholderBuilder(
-                {'lib.txt': 'libText'}.build(),
-                inputPlaceholder: r'$lib$',
-              ),
-            ),
-            applyToRoot(
-              PlaceholderBuilder(
-                {'root.txt': 'rootText'}.build(),
-                inputPlaceholder: r'$package$',
-              ),
-            ),
-          ],
+        final builder1 = PlaceholderBuilder(
+          {'lib.txt': 'libText'}.build(),
+          inputPlaceholder: r'$lib$',
+        );
+        final builder2 = PlaceholderBuilder(
+          {'root.txt': 'rootText'}.build(),
+          inputPlaceholder: r'$package$',
+        );
+        await testBuilders(
+          [builder1, builder2],
           {},
+          visibleOutputBuilders: {builder1, builder2},
+          rootPackage: 'a',
           outputs: {'a|lib/lib.txt': 'libText', 'a|root.txt': 'rootText'},
         );
       });
@@ -1030,16 +1027,12 @@ targets:
     });
 
     test('can\'t read files in .dart_tool', () async {
-      await testPhases(
-        [
-          apply('', [
-            (_) => TestBuilder(
-              build: copyFrom(makeAssetId('a|.dart_tool/any_file')),
-            ),
-          ], toRoot()),
-        ],
-        {'a|lib/a.txt': 'a', 'a|.dart_tool/any_file': 'content'},
-        status: BuildStatus.failure,
+      expect(
+        (await testBuilders(
+          [TestBuilder(build: copyFrom(makeAssetId('a|.dart_tool/any_file')))],
+          {'a|lib/a.txt': 'a', 'a|.dart_tool/any_file': 'content'},
+        )).buildResult.status,
+        BuildStatus.failure,
       );
     });